转移收藏相关操作到收藏页面
